move functional tests
