+++ functional_tests/steps/breadcrumb.py
@@ -1,0 +1,13 @@
+from behave import then  # pylint: disable=no-name-in-module
+from behave.runner import Context
+from playwright.sync_api import expect
+
+
+@then("they cannot see the breadcrumbs")
+def user_does_not_sees_on_the_page(context: Context) -> None:
+    expect(context.page.get_by_label("Breadcrumbs")).not_to_be_visible()
+
+
+@then("the user can see the breadcrumbs")
+def user_does_sees_on_the_page(context: Context) -> None:
+    expect(context.page.get_by_label("Breadcrumbs")).to_be_visible()
